Move mock app fixturing to mock app module
